Change VitePress config to add locales
